Move control snap helpers to processorhelpers
